Rearrange processors
