+++ src/input_processing/dual_axis/circle.rs
@@ -1,4 +1,4 @@
-//! Circular input boundaries for dual-axis inputs
+//! Circular range processors for dual-axis inputs
 
 use std::fmt::Debug;
 use std::hash::{Hash, Hasher};
