+++ src/input_processing/single_axis/range.rs
@@ -1,249 +1,11 @@
-//! Processors for single-axis input values
+//! Range processors for single-axis inputs
 
-use std::fmt::Debug;
+use bevy::prelude::Reflect;
+use bevy::utils::FloatOrd;
+use serde::{Deserialize, Serialize};
 use std::hash::{Hash, Hasher};
 
-use bevy::prelude::*;
-use bevy::utils::FloatOrd;
-use dyn_clone::DynClone;
-use dyn_eq::DynEq;
-use dyn_hash::DynHash;
-use serde::{Deserialize, Serialize};
-
-/// A trait for processing single-axis input values,
-/// accepting a `f32` input and producing a `f32` output.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::hash::{Hash, Hasher};
-/// use bevy::prelude::*;
-/// use bevy::utils::FloatOrd;
-/// use serde::{Deserialize, Serialize};
-/// use leafwing_input_manager::prelude::*;
-///
-/// /// Doubles the input, takes the absolute value,
-/// /// and discards results that meet the specified condition.
-/// // If your processor includes fields not implemented Eq and Hash,
-/// // implementation is necessary as shown below.
-/// // Otherwise, you can derive Eq and Hash directly.
-/// #[derive(Debug, Clone, PartialEq, Reflect, Serialize, Deserialize)]
-/// pub struct DoubleAbsoluteValueThenIgnored(pub f32);
-///
-/// // Add this attribute for ensuring proper serialization and deserialization.
-/// #[serde_trait_object]
-/// impl AxisProcessor for DoubleAbsoluteValueThenIgnored {
-///     fn process(&self, input_value: f32) -> f32 {
-///         // Implement the logic just like you would in a normal function.
-///
-///         // You can use other processors within this function.
-///         let value = AxisSensitivity(2.0).process(input_value);
-///
-///         let value = value.abs();
-///         if value == self.0 {
-///             0.0
-///         } else {
-///             value
-///         }
-///     }
-/// }
-///
-/// // Unfortunately, manual implementation is required due to the float field.
-/// impl Eq for DoubleAbsoluteValueThenIgnored {}
-/// impl Hash for DoubleAbsoluteValueThenIgnored {
-///     fn hash<H: Hasher>(&self, state: &mut H) {
-///         // Encapsulate the float field for hashing.
-///         FloatOrd(self.0).hash(state);
-///     }
-/// }
-///
-/// // Now you can use it!
-/// let processor = DoubleAbsoluteValueThenIgnored(4.0);
-///
-/// // Rejected!
-/// assert_eq!(processor.process(2.0), 0.0);
-/// assert_eq!(processor.process(-2.0), 0.0);
-///
-/// // Others are just doubled absolute value.
-/// assert_eq!(processor.process(6.0), 12.0);
-/// assert_eq!(processor.process(4.0), 8.0);
-/// assert_eq!(processor.process(0.0), 0.0);
-/// assert_eq!(processor.process(-4.0), 8.0);
-/// assert_eq!(processor.process(-6.0), 12.0);
-/// ```
-#[typetag::serde(tag = "type")]
-pub trait AxisProcessor: Send + Sync + Debug + DynClone + DynEq + DynHash + Reflect {
-    /// Computes the result by processing the `input_value`.
-    fn process(&self, input_value: f32) -> f32;
-}
-
-dyn_clone::clone_trait_object!(AxisProcessor);
-dyn_eq::eq_trait_object!(AxisProcessor);
-dyn_hash::hash_trait_object!(AxisProcessor);
-crate::__reflect_trait_object!(AxisProcessor);
-
-/// A dynamic sequence container of [`AxisProcessor`]s designed for processing input values.
-///
-/// # Warning
-///
-/// This flexibility may hinder compiler optimizations such as inlining or dead code elimination.
-/// For performance-critical scenarios, consider creating your own processors for improved performance.
-///
-/// # Examples
-///
-/// ```rust
-/// use leafwing_input_manager::prelude::*;
-///
-/// // Just a heads up, the default pipeline won't tweak values.
-/// let pipeline = AxisProcessingPipeline::default();
-/// assert_eq!(pipeline.process(1.5), 1.5);
-///
-/// // You can link up a sequence of processors to make a pipeline.
-/// let mut pipeline = AxisProcessingPipeline::default()
-///     .with(AxisSensitivity(2.0))
-///     .with(AxisInverted);
-///
-/// // Now it doubles and flips values.
-/// assert_eq!(pipeline.process(1.5), -3.0);
-///
-/// // You can also add processors just like you would do with a Vec.
-/// pipeline.push(AxisSensitivity(1.5));
-///
-/// // Now it triples and inverts values.
-/// assert_eq!(pipeline.process(1.5), -4.5);
-///
-/// // Plus, you can switch out a processor at a specific index.
-/// pipeline.set(2, AxisSensitivity(-2.0));
-///
-/// // Now it multiplies values by -4 and inverts the result.
-/// assert_eq!(pipeline.process(1.5), 6.0);
-///
-/// // If needed, you can clear out all processors.
-/// pipeline.clear();
-///
-/// // Now it just leaves values as is.
-/// assert_eq!(pipeline.process(1.5), 1.5);
-/// ```
-#[must_use]
-#[derive(Debug, Default, Clone, PartialEq, Eq, Hash, Reflect, Serialize, Deserialize)]
-pub struct AxisProcessingPipeline(pub(crate) Vec<Box<dyn AxisProcessor>>);
-
-#[typetag::serde]
-impl AxisProcessor for AxisProcessingPipeline {
-    /// Computes the result by passing the `input_value` through this pipeline.
-    #[must_use]
-    #[inline]
-    fn process(&self, input_value: f32) -> f32 {
-        self.0
-            .iter()
-            .fold(input_value, |value, next| next.process(value))
-    }
-}
-
-impl AxisProcessingPipeline {
-    /// Appends the given [`AxisProcessor`] into this pipeline and returns `self`.
-    #[inline]
-    pub fn with(mut self, processor: impl AxisProcessor) -> Self {
-        self.push(processor);
-        self
-    }
-
-    /// Appends the given [`AxisProcessor`] into this pipeline.
-    #[inline]
-    pub fn push(&mut self, processor: impl AxisProcessor) {
-        self.0.push(Box::new(processor));
-    }
-
-    /// Replaces the processor at the `index` with the given [`AxisProcessor`].
-    #[inline]
-    pub fn set(&mut self, index: usize, processor: impl AxisProcessor) {
-        self.0[index] = Box::new(processor);
-    }
-
-    /// Removes all processors in this pipeline.
-    #[inline]
-    pub fn clear(&mut self) {
-        self.0.clear();
-    }
-}
-
-/// A trait for appending an [`AxisProcessor`] as the next processing step in the pipeline,
-/// enabling further processing of input values.
-pub trait WithNextAxisProcessor {
-    /// Appends an [`AxisProcessor`] as the next processing step in the pipeline.
-    fn with_processor(self, processor: impl AxisProcessor) -> Self;
-}
-
-impl WithNextAxisProcessor for Box<dyn AxisProcessor> {
-    /// Creates a new boxed [`AxisProcessingPipeline`] with the existing steps
-    /// and appends the given [`AxisProcessor`].
-    fn with_processor(self, processor: impl AxisProcessor) -> Self {
-        let pipeline = match Reflect::as_any(&*self).downcast_ref::<AxisProcessingPipeline>() {
-            Some(pipeline) => pipeline.clone(),
-            None => AxisProcessingPipeline(vec![self]),
-        };
-        Box::new(pipeline.with(processor))
-    }
-}
-
-/// Flips the sign of single-axis input values, resulting in a directional reversal of control.
-///
-/// ```rust
-/// use leafwing_input_manager::prelude::*;
-///
-/// assert_eq!(AxisInverted.process(2.5), -2.5);
-/// assert_eq!(AxisInverted.process(-2.5), 2.5);
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Reflect, Serialize, Deserialize)]
-#[must_use]
-pub struct AxisInverted;
-
-#[typetag::serde]
-impl AxisProcessor for AxisInverted {
-    /// Returns the opposite value of the `input_value`.
-    #[must_use]
-    #[inline]
-    fn process(&self, input_value: f32) -> f32 {
-        -input_value
-    }
-}
-
-/// Scales input values on an axis using a specified multiplier,
-/// allowing fine-tuning the responsiveness of controls.
-///
-/// ```rust
-/// use leafwing_input_manager::prelude::*;
-///
-/// // Doubled!
-/// assert_eq!(AxisSensitivity(2.0).process(2.0), 4.0);
-///
-/// // Halved!
-/// assert_eq!(AxisSensitivity(0.5).process(2.0), 1.0);
-///
-/// // Negated and halved!
-/// assert_eq!(AxisSensitivity(-0.5).process(2.0), -1.0);
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Reflect, Serialize, Deserialize)]
-#[must_use]
-pub struct AxisSensitivity(pub f32);
-
-#[typetag::serde]
-impl AxisProcessor for AxisSensitivity {
-    /// Multiples the `input_value` by the specified sensitivity factor.
-    #[must_use]
-    #[inline]
-    fn process(&self, input_value: f32) -> f32 {
-        self.0 * input_value
-    }
-}
-
-impl Eq for AxisSensitivity {}
-
-impl Hash for AxisSensitivity {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        FloatOrd(self.0).hash(state);
-    }
-}
+use super::AxisProcessor;
 
 /// Specifies an acceptable min-max range for valid single-axis inputs,
 /// restricting all value stays within intended limits
@@ -751,74 +513,8 @@ impl Hash for AxisDeadZone {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
-    #[test]
-    fn test_axis_processing_pipeline() {
-        // Chain processors to make a new pipeline.
-        let mut pipeline = AxisProcessingPipeline::default()
-            .with(AxisSensitivity(4.0))
-            .with(AxisInverted)
-            .with(AxisSensitivity(4.0));
-
-        pipeline.push(AxisInverted);
-
-        pipeline.set(2, AxisSensitivity(6.0));
-
-        // This pipeline now scales input values by a factor of 24.0
-        assert_eq!(pipeline.process(2.0), 48.0);
-        assert_eq!(pipeline.process(-3.0), -72.0);
-
-        // Now it just leaves values as is.
-        pipeline.clear();
-        assert_eq!(pipeline, AxisProcessingPipeline::default());
-        assert_eq!(pipeline.process(4.0), 4.0);
-    }
-
-    #[test]
-    fn test_merge_axis_processor() {
-        let first = AxisSensitivity(2.0);
-        let first_boxed: Box<dyn AxisProcessor> = Box::new(first);
-
-        let second = AxisSensitivity(3.0);
-        let merged_second = first_boxed.with_processor(second);
-        let expected = AxisProcessingPipeline::default().with(first).with(second);
-        let expected_boxed: Box<dyn AxisProcessor> = Box::new(expected);
-        assert_eq!(merged_second, expected_boxed);
-
-        let third = AxisSensitivity(4.0);
-        let merged_third = merged_second.with_processor(third);
-        let expected = AxisProcessingPipeline::default()
-            .with(first)
-            .with(second)
-            .with(third);
-        let expected_boxed: Box<dyn AxisProcessor> = Box::new(expected);
-        assert_eq!(merged_third, expected_boxed);
-    }
-
-    #[test]
-    fn test_axis_inverted() {
-        for value in -300..300 {
-            let value = value as f32 * 0.01;
-
-            assert_eq!(AxisInverted.process(value), -value);
-            assert_eq!(AxisInverted.process(-value), value);
-        }
-    }
-
-    #[test]
-    fn test_axis_sensitivity() {
-        for value in -300..300 {
-            let value = value as f32 * 0.01;
-
-            for sensitivity in -300..300 {
-                let sensitivity = sensitivity as f32 * 0.01;
-
-                let scale = AxisSensitivity(sensitivity);
-                assert_eq!(scale.process(value), sensitivity * value);
-            }
-        }
-    }
+    use crate::input_processing::*;
+    use bevy::prelude::FloatExt;
 
     #[test]
     fn test_axis_value_bounds() {
